Identity added

+++ src/AuctionService-Controllers/Controllers/AuctionController.cs
@@ -3,6 +3,9 @@ using AuctionService_Controllers.Dtos;
 using AuctionService_Controllers.Entities;
 using AutoMapper;
 using AutoMapper.QueryableExtensions;
+using Contracts;
+using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -14,11 +17,16 @@ namespace AuctionService_Controllers.Controllers
     {
         private readonly AuctionDbContext _context;
         private readonly IMapper _mapper;
+        private readonly IPublishEndpoint _publishEndpoint;
 
-        public AuctionController(AuctionDbContext context, IMapper mapper)
+
+        public AuctionController(AuctionDbContext context, IMapper mapper,
+        IPublishEndpoint publishEndpoint)
         {
             _context = context;
             _mapper = mapper;
+            _publishEndpoint = publishEndpoint;
+
         }
 
         [HttpGet]
@@ -28,7 +36,7 @@ namespace AuctionService_Controllers.Controllers
 
             if (!string.IsNullOrEmpty(date))
             {
-                query = query.Where(x => x.UpdatedAt.CompareTo(DateTime.Parse(date)) > 0);
+                query = query.Where(x => x.UpdatedAt.CompareTo(DateTime.Parse(date).ToUniversalTime()) > 0);
             }
 
             return query.ProjectTo<AuctionDto>(_mapper.ConfigurationProvider).ToList();
@@ -46,20 +54,28 @@ namespace AuctionService_Controllers.Controllers
             return _mapper.Map<AuctionDto>(auction);
         }
 
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDto>> CreateAuction(CreateActionDto auctionDto)
         {
             var auction = _mapper.Map<Auction>(auctionDto);
-            auction.Seller = "test";
+
+            auction.Seller = User.Identity.Name;
 
             _context.Auctions.Add(auction);
 
+            var newAuction = _mapper.Map<AuctionDto>(auction);
+
+            await _publishEndpoint.Publish(_mapper.Map<AuctionCreated>(newAuction));
+
             var result = await _context.SaveChangesAsync() > 0;
+
             if (!result) return BadRequest("Coudn't save changes to DB");
 
-            return CreatedAtAction(nameof(GetAuctionById), new { auction.Id }, _mapper.Map<AuctionDto>(auction));
+            return CreatedAtAction(nameof(GetAuctionById), new { auction.Id }, newAuction);
         }
 
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
         {
@@ -68,13 +84,16 @@ namespace AuctionService_Controllers.Controllers
 
             if (auction == null) return NotFound();
 
-            // TODO: check seller == username
+            if (auction.Seller != User?.Identity?.Name)
+                return Forbid();
 
             auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
             auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
             auction.Item.Color = updateAuctionDto.Color ?? auction.Item.Color;
             auction.Item.Mileage = updateAuctionDto.Mileage ?? auction.Item.Mileage;
             auction.Item.Year = updateAuctionDto.Year ?? auction.Item.Year;
+
+            await _publishEndpoint.Publish(_mapper.Map<AuctionUpdated>(auction));
 
             var result = await _context.SaveChangesAsync() > 0;
 
@@ -83,6 +102,7 @@ namespace AuctionService_Controllers.Controllers
             return BadRequest("Problem saving changes");
         }
 
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteAuction(Guid id)
         {
@@ -90,9 +110,12 @@ namespace AuctionService_Controllers.Controllers
 
             if (auction == null) return NotFound();
 
-            // TODO: check seller == username
+            if (auction.Seller != User?.Identity?.Name)
+                return Forbid();
 
             _context.Auctions.Remove(auction);
+
+            await _publishEndpoint.Publish<AuctionDeleted>(new { Id = auction.Id.ToString() });
 
             var result = await _context.SaveChangesAsync() > 0;
